msm: cpp: Add error handling for cpp poll and load firmware functions
If CPP micro is in bad state, driver reloads the firmware and
reschedules the payload. This helps in recovery from most error
conditions. During stability runs if hardware is in bad state,
reload firmware and rescheduling the payload fails. In these cases
error is not handled in poll/ load firmware API causing frames to
be send to hardware in bad state. This can cause stability failures
and device crashes. Add error handling for cpp poll and load firmware
functions and prevent sending frames to hardware for processing if
hardware is in bad state.

CRs-Fixed: 876275
Change-Id: I85b1130959d941ad49d6c90c24b1d284cfc116b2
Signed-off-by: Krishnankutty Kolathappilly <kkolatha@codeaurora.org>
Signed-off-by: Sunid Wilson <sunidw@codeaurora.org>