msm: vmem: Temporary workaround for vmem allocate on V3
On 8996 V3, vmem clock is is left turned ON in hardware,
even though HLOS tried to disable the clock during vmem probe.
Because of this vmem_allocate() would fail during init as the
vmem bank is not in sleep state.
UBWC encode mandates vmem to be allocated and set to firmware,
else we see corruption in encoded output.
This change is a temporary workaround to fix the issue...

Change-Id: Ie56cb431959ccb2daf6ad15048b182998636d0b8
Signed-off-by: Sowmya Pandiri <spandiri@codeaurora.org>