msm_defconfig: cleanup defconfig
Generate a minimized defconfig by running "make menuconfig" and
saving, without actually making any changes. This strips out
redundant lines that specify configurations which are default or
implicit based on other options, and reduces the potential for
conflicts when future changes to the defconfig are made.

As an added bonus, this avoids diffs appearing in the defconfig
as a result of simply building the kernel.

No functional change is expected due to this change.

Change-Id: I374e06200cb5e443b676d9274bf5dda405aa8ead
Signed-off-by: Scott Mertz <scott@cyngn.com>