defconfig: remove msm-perf_ailsa_ii_defconfig
We are using the default msm_defconfig for the ailsa_ii project.
Remove the other config to eliminate confusion.

Note: At a future time, we need to move the debug configs from
msm_defconfig to a new defconfig snippet that can be used with
KERNEL_ADDITIONAL_CONFIG (in merge-config.sh) instead.

Change-Id: I77a04129906223697a19e979d9c940ea7c174b9a